--- a/common/crypto/bn254/cloudflare/bn254_test.go
+++ b/common/crypto/bn254/cloudflare/bn254_test.go
@@ -94,23 +94,21 @@ func TestTripartiteDiffieHellman(t *testing.T) {
 
 func BenchmarkG1(b *testing.B) {
 	x, _ := rand.Int(rand.Reader, Order)
-	b.ResetTimer()
 
-	for i := 0; i < b.N; i++ {
+	for b.Loop() {
 		new(G1).ScalarBaseMult(x)
 	}
 }
 
 func BenchmarkG2(b *testing.B) {
 	x, _ := rand.Int(rand.Reader, Order)
-	b.ResetTimer()
 
-	for i := 0; i < b.N; i++ {
+	for b.Loop() {
 		new(G2).ScalarBaseMult(x)
 	}
 }
 func BenchmarkPairing(b *testing.B) {
-	for i := 0; i < b.N; i++ {
+	for b.Loop() {
 		Pair(&G1{curveGen}, &G2{twistGen})
 	}
 }