--- a/common/crypto/crypto_test.go
+++ b/common/crypto/crypto_test.go
@@ -95,7 +95,7 @@ func TestToECDSAErrors(t *testing.T) {
 
 func BenchmarkSha3(b *testing.B) {
 	a := []byte("hello world")
-	for i := 0; i < b.N; i++ {
+	for b.Loop() {
 		Keccak256(a)
 	}
 }