These changes use b.Loop() to simplify the code and improve performance

Supported by Go Team, more info: https://go.dev/blog/testing-b-loop

More info can see https://go.dev/issue/73137.

Before:

```shell
go test -run=^$ -bench=. ./common/crypto -timeout=1h     
goos: darwin
goarch: arm64
pkg: github.com/erigontech/erigon/common/crypto
cpu: Apple M4
BenchmarkSha3-10                  	 4661552	       240.0 ns/op
BenchmarkEcrecoverSignature-10    	   74696	     16256 ns/op
BenchmarkVerifySignature-10       	   94023	     13391 ns/op
BenchmarkDecompressPubkey-10      	  526609	      2280 ns/op
PASS
ok  	github.com/erigontech/erigon/common/crypto	5.807s

```

After:

```shell
go test -run=^$ -bench=. ./common/crypto -timeout=1h  
goos: darwin
goarch: arm64
pkg: github.com/erigontech/erigon/common/crypto
cpu: Apple M4
BenchmarkSha3-10                  	 4623690	       253.0 ns/op
BenchmarkEcrecoverSignature-10    	   77833	     15569 ns/op
BenchmarkVerifySignature-10       	   91219	     13073 ns/op
BenchmarkDecompressPubkey-10      	  526153	      2268 ns/op
PASS
ok  	github.com/erigontech/erigon/common/crypto	5.181s
```